reorg before implementing varint encoding
