+++ borsh/Cargo.toml
@@ -21,8 +21,8 @@ name = "generate_schema_schema"
 path = "src/generate_schema_schema.rs"
 
 [dependencies]
-bigdecimal = { version = "0.3", optional = true }
-num-bigint = { version = "0.4.3", optional = true }
+bigdecimal-dep = { package = "bigdecimal", version = "0.3", optional = true }
+num-bigint-dep = { package = "num-bigint", version = "0.4.3", optional = true }
 borsh-derive = { path = "../borsh-derive" }
 hashbrown = "0.11"
 
@@ -33,4 +33,6 @@ quickcheck = "1.0"
 default = ["std"]
 std = []
 rc = []
+bigdecimal = ["num-bigint"]
+num-bigint = ["num-bigint-dep"]
 const-generics = []
