+++ borsh/src/lib.rs
@@ -10,6 +10,9 @@ pub mod schema;
 pub mod schema_helpers;
 pub mod ser;
 
+#[cfg(feature = "num-bigint")]
+mod bigint;
+
 pub use de::BorshDeserialize;
 pub use schema::BorshSchema;
 pub use schema_helpers::{try_from_slice_with_schema, try_to_vec_with_schema};
